+++ changelog.md
@@ -42,6 +42,7 @@
 - Added class `Projects/Secrets`
 - Added class `Projects/Secrets/DefaultBundle`
 - Added class `Projects/Tags`
+- Added class `Projects/Vault`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`